done should close the accordion
