+++ frontend/app/components/redesign/components/AppearanceBuilder.tsx
@@ -65,7 +65,7 @@ export type ToolAppearance = BannerToolAppearance | WidgetToolAppearance
 interface AppearanceBuilderProps {
   appearance: ToolAppearance
   isComplete?: boolean
-  onDone?: () => void
+  onDone: () => void
   positionSelector?: React.ReactNode
   colorsSelector?: React.ReactNode
   activeVersion?: string
@@ -114,12 +114,6 @@ export const AppearanceBuilder: React.FC<AppearanceBuilderProps> = ({
     console.log('Refresh')
   }
 
-  const handleDoneClick = () => {
-    if (onDone) {
-      onDone()
-    }
-  }
-
   return (
     <BuilderAccordion
       title="Appearance"
@@ -127,6 +121,7 @@ export const AppearanceBuilder: React.FC<AppearanceBuilderProps> = ({
       activeVersion={activeVersion}
       onToggle={handleToggle}
       onRefresh={handleRefresh}
+      onDone={onDone}
     >
       <div className="flex flex-col gap-xs">
         <SectionHeader icon={<SVGText className="w-5 h-5" />} label="Text" />
@@ -285,17 +280,6 @@ export const AppearanceBuilder: React.FC<AppearanceBuilderProps> = ({
             ))}
           </div>
         </div>
-      </div>
-
-      <Divider />
-
-      <div className="flex justify-end">
-        <ToolsSecondaryButton
-          className="w-full xl:w-[140px]"
-          onClick={handleDoneClick}
-        >
-          Done
-        </ToolsSecondaryButton>
       </div>
     </BuilderAccordion>
   )
